requirements.txt
Flask>=3.0
Flask-SQLAlchemy>=3.1
python-dotenv>=1.0
bcrypt>=4.0
waitress>=3.0
pandas>=2.0
